requirements.txt
azure-search-documents
chainlit==1.3.0rc1
openai
yfinance
plotly
websockets
pydantic==2.10.1
azure-identity==1.19.0
azure-cognitiveservices-speech
aiohttp
torch
torchaudio